Remove unnecessary afw::table Writer base class.
